[백영주] 관리자페이지(회원목록생성) - 세부사항 	댓글 수정삭제는 작성자만 가능하도록 권한설정 	회원탈퇴시 작성한 가계부와 댓글 삭제되도록 구현 	관리자로그인시 사이드메뉴에 회원목록버튼 활성화 	관리자가 아닐시 비활성화

+++ cashbook/src/main/java/com/gdu/cashbook/controller/CommentController.java
@@ -53,9 +53,16 @@ public class CommentController {
 	public String removeComment(HttpSession session,
 			Model model,
 			@RequestParam(value = "commentNo", required = false, defaultValue = "0")int commentNo) {
+		if(commentNo == 0) {
+			return "redirect:/";
+		}
+		LoginMember loginMember = (LoginMember)session.getAttribute("loginMember");
 		Comment comment = commentService.getComment(commentNo);
+		if(loginMember.getMemberId().equals(comment.getMemberId())) {
+			commentService.removeComment(commentNo);
+		}
 		Board board = boardService.getBoardOne(comment.getBoardNo());
-		commentService.removeComment(commentNo);
+		
 		model.addAttribute("board", board);
 		int rowPerPage=10;
 		
@@ -86,6 +93,8 @@ public class CommentController {
 			Model model,
 			@RequestParam(value = "commentNo", required = false, defaultValue = "0")int commentNo) {
 		Comment comment = commentService.getComment(commentNo);
+		LoginMember loginMember = (LoginMember)session.getAttribute("loginMember");
+		
 		Board board = boardService.getBoardOne(comment.getBoardNo());
 		model.addAttribute("board", board);
 		
@@ -96,7 +105,9 @@ public class CommentController {
 		
 		//commentNo for modify
 		model.addAttribute("comment", comment);
-		
+		if(loginMember.getMemberId().equals(comment.getMemberId())) {
 		return "modifyComment";
+		}
+		return "boardOneList";
 	}
 }
